requirements.txt
python-telegram-bot==20.4
ApScheduler==3.10.1
Django==4.2.4
python-dotenv=1.0.0